Change the data type of the receive arrays for global can

+++ GlobalCAN_IDs.h
@@ -7,7 +7,7 @@
 
 #ifndef GLOBALCAN_IDS_H
 #define	GLOBALCAN_IDS_H
-
+#include <stdint.h>
 #ifdef	__cplusplus
 extern "C" {
 #endif
@@ -49,7 +49,7 @@ typedef enum {
 #define getGBL_DEVICE_STATUS(c) c*GLOBAL_DATA_IDX_PER_DEV + DEVICE_STATUS
 #define getGBL_MACRO_INDEX(c) c*GLOBAL_DATA_IDX_PER_DEV + DEVICE_MACRO
 
-volatile int receiveArrayCAN_Global[GLOBAL_DEVICES*GLOBAL_DATA_IDX_PER_DEV +1];
+volatile short receiveArrayCAN_Global[GLOBAL_DEVICES*GLOBAL_DATA_IDX_PER_DEV +1];
 volatile bool GBL_CAN_FT_recievedFlag[GLOBAL_DEVICES*GLOBAL_DATA_IDX_PER_DEV +1];
 
 #ifdef	__cplusplus
